requirements.txt
python-whois
openai
python-dotenv